+++ configs/milvus.yaml
@@ -41,7 +41,7 @@ etcd:
     enabled: false # Whether to support ETCD secure connection mode
     tlsCert: /path/to/etcd-client.pem # path to your cert file
     tlsKey: /path/to/etcd-client-key.pem # path to your key file
-    tlsCACert:  /path/to/ca.pem # path to your CACert file
+    tlsCACert: /path/to/ca.pem # path to your CACert file
     # TLS min version
     # Optional values: 1.0, 1.1, 1.2, 1.3。
     # We recommend using version 1.2 and above
@@ -114,7 +114,7 @@ rocksmq:
   retentionTimeInMinutes: 7200 # 5 days, 5 * 24 * 60 minutes, The retention time of the message in rocksmq.
   retentionSizeInMB: 8192 # 8 GB, 8 * 1024 MB, The retention size of the message in rocksmq.
   compactionInterval: 86400 # 1 day, trigger rocksdb compaction every day to remove deleted data
-  lrucacheratio:  0.06 # rocksdb cache memory ratio
+  lrucacheratio: 0.06 # rocksdb cache memory ratio
 
 # Related configuration of rootCoord, used to handle data definition language (DDL) and data control language (DCL) requests
 rootCoord:
@@ -175,14 +175,11 @@ queryCoord:
 
 # Related configuration of queryNode, used to run hybrid search between vector and scalar data.
 queryNode:
-  cacheSize: 32 # GB, default 32 GB, `cacheSize` is the memory used for caching data for faster query. The `cacheSize` must be less than system memory size.
   port: 21123
   loadMemoryUsageFactor: 3 # The multiply factor of calculating the memory usage while loading segments
   enableDisk: true # enable querynode load disk index, and search on disk index
   maxDiskUsagePercentage: 95
 
-  stats:
-    publishInterval: 1000 # Interval for querynode to report node information (milliseconds)
   dataSync:
     flowGraph:
       maxQueueLength: 1024 # Maximum length of task queue in flowgraph
@@ -245,7 +242,7 @@ dataCoord:
 
   segment:
     maxSize: 512 # Maximum size of a segment in MB
-    diskSegmentMaxSize:  2048 # Maximun size of a segment in MB for collection which has Disk index
+    diskSegmentMaxSize: 2048 # Maximun size of a segment in MB for collection which has Disk index
     sealProportion: 0.25 # It's the minimum proportion for a segment which can be sealed
     assignmentExpiration: 2000 # The time of the assignment expiration in ms
     maxLife: 86400 # The max lifetime of segment in seconds, 24*60*60
@@ -295,8 +292,8 @@ grpc:
   clientMaxSendSize: 104857600 # 100 MB, 100 * 1024 * 1024
 
   client:
-    dialTimeout:      5000
-    keepAliveTime:    10000
+    dialTimeout: 5000
+    keepAliveTime: 10000
     keepAliveTimeout: 20000
     maxMaxAttempts: 5
     initialBackOff: 1.0
@@ -318,10 +315,7 @@ common:
     rootCoordStatistics: "rootcoord-statistics"
     rootCoordDml: "rootcoord-dml"
     rootCoordDelta: "rootcoord-delta"
-    search: "search"
-    searchResult: "searchResult"
     queryTimeTick: "queryTimeTick"
-    queryNodeStats: "query-node-stats"
     # Cmd for loadIndex, flush, etc...
     cmd: "cmd"
     dataCoordStatistic: "datacoord-statistics-channel"
@@ -339,7 +333,7 @@ common:
   defaultPartitionName: "_default"  # default partition name for a collection
   defaultIndexName: "_default_idx"  # default index name
   retentionDuration: 86400 # time travel reserved time, insert/delete will not be cleaned in this period. 1 days in seconds
-  entityExpiration:  -1     # Entity expiration in seconds, CAUTION make sure entityExpiration >= retentionDuration and -1 means never expire
+  entityExpiration: -1     # Entity expiration in seconds, CAUTION make sure entityExpiration >= retentionDuration and -1 means never expire
 
   gracefulTime: 5000 # milliseconds. it represents the interval (in ms) by which the request arrival time needs to be subtracted in the case of Bounded Consistency.
 
@@ -452,3 +446,7 @@ quotaAndLimits:
 
       # coolOffSpeed is the speed of search&query rates cool off.
       coolOffSpeed: 0.9 # (0, 1]
+
+# AutoIndexConfig
+autoIndex:
+  enable: false
