+++ internal/metastore/db/dbmodel/index.go
@@ -14,18 +14,20 @@ import (
 )
 
 type Index struct {
-	ID           int64     `gorm:"id"`
-	TenantID     string    `gorm:"tenant_id"`
-	FieldID      int64     `gorm:"field_id"`
-	CollectionID int64     `gorm:"collection_id"`
-	IndexID      int64     `gorm:"index_id"`
-	IndexName    string    `gorm:"index_name"`
-	IndexParams  string    `gorm:"index_params"`
-	TypeParams   string    `gorm:"type_params"`
-	CreateTime   uint64    `gorm:"create_time"`
-	IsDeleted    bool      `gorm:"is_deleted"`
-	CreatedAt    time.Time `gorm:"created_at"`
-	UpdatedAt    time.Time `gorm:"updated_at"`
+	ID              int64     `gorm:"id"`
+	TenantID        string    `gorm:"tenant_id"`
+	FieldID         int64     `gorm:"field_id"`
+	CollectionID    int64     `gorm:"collection_id"`
+	IndexID         int64     `gorm:"index_id"`
+	IndexName       string    `gorm:"index_name"`
+	IndexParams     string    `gorm:"index_params"`
+	TypeParams      string    `gorm:"type_params"`
+	CreateTime      uint64    `gorm:"create_time"`
+	IsDeleted       bool      `gorm:"is_deleted"`
+	CreatedAt       time.Time `gorm:"created_at"`
+	UpdatedAt       time.Time `gorm:"updated_at"`
+	IsAutoIndex     bool      `gorm:"is_auto_index"`
+	UserIndexParams string    `gorm:"index_params"`
 }
 
 func (v Index) TableName() string {
@@ -35,14 +37,16 @@ func (v Index) TableName() string {
 // ------------- search result -------------
 
 type IndexResult struct {
-	FieldID      int64
-	CollectionID int64
-	IndexID      int64
-	IndexName    string
-	TypeParams   string
-	IndexParams  string
-	CreateTime   uint64
-	IsDeleted    bool
+	FieldID         int64
+	CollectionID    int64
+	IndexID         int64
+	IndexName       string
+	TypeParams      string
+	IndexParams     string
+	CreateTime      uint64
+	IsDeleted       bool
+	IsAutoIndex     bool
+	UserIndexParams string
 }
 
 //go:generate mockery --name=IIndexDb
@@ -69,6 +73,15 @@ func UnmarshalIndexModel(inputs []*IndexResult) ([]*model.Index, error) {
 				return nil, err
 			}
 		}
+		var userIndexParams []commonpb.KeyValuePair
+		if ir.UserIndexParams != "" {
+			err := json.Unmarshal([]byte(ir.UserIndexParams), &userIndexParams)
+			if err != nil {
+				log.Error("unmarshal UserIndexParams of index failed", zap.Int64("collID", ir.CollectionID),
+					zap.Int64("indexID", ir.IndexID), zap.String("indexName", ir.IndexName), zap.Error(err))
+				return nil, err
+			}
+		}
 
 		var typeParams []commonpb.KeyValuePair
 		if ir.TypeParams != "" {
@@ -81,14 +94,16 @@ func UnmarshalIndexModel(inputs []*IndexResult) ([]*model.Index, error) {
 		}
 
 		idx := &model.Index{
-			CollectionID: ir.CollectionID,
-			FieldID:      ir.FieldID,
-			IndexID:      ir.IndexID,
-			IndexName:    ir.IndexName,
-			IndexParams:  funcutil.ConvertToKeyValuePairPointer(indexParams),
-			TypeParams:   funcutil.ConvertToKeyValuePairPointer(typeParams),
-			CreateTime:   ir.CreateTime,
-			IsDeleted:    ir.IsDeleted,
+			CollectionID:    ir.CollectionID,
+			FieldID:         ir.FieldID,
+			IndexID:         ir.IndexID,
+			IndexName:       ir.IndexName,
+			IndexParams:     funcutil.ConvertToKeyValuePairPointer(indexParams),
+			TypeParams:      funcutil.ConvertToKeyValuePairPointer(typeParams),
+			CreateTime:      ir.CreateTime,
+			IsDeleted:       ir.IsDeleted,
+			IsAutoIndex:     ir.IsAutoIndex,
+			UserIndexParams: funcutil.ConvertToKeyValuePairPointer(userIndexParams),
 		}
 		result = append(result, idx)
 	}
