+++ internal/metastore/db/indexcoord/table_catalog.go
@@ -36,6 +36,14 @@ func (tc *Catalog) CreateIndex(ctx context.Context, index *model.Index) error {
 		return err
 	}
 
+	userIndexParamsBytes, err := json.Marshal(index.UserIndexParams)
+	if err != nil {
+		log.Error("marshal IndexParams of index failed", zap.String("tenant", tenantID),
+			zap.Int64("collID", index.CollectionID), zap.Int64("indexID", index.IndexID),
+			zap.String("indexName", index.IndexName), zap.Error(err))
+		return err
+	}
+
 	typeParamsBytes, err := json.Marshal(index.TypeParams)
 	if err != nil {
 		log.Error("marshal TypeParams of index failed", zap.String("tenant", tenantID),
@@ -45,15 +53,17 @@ func (tc *Catalog) CreateIndex(ctx context.Context, index *model.Index) error {
 	}
 
 	idx := &dbmodel.Index{
-		TenantID:     tenantID,
-		CollectionID: index.CollectionID,
-		FieldID:      index.FieldID,
-		IndexID:      index.IndexID,
-		IndexName:    index.IndexName,
-		TypeParams:   string(typeParamsBytes),
-		IndexParams:  string(indexParamsBytes),
-		CreateTime:   index.CreateTime,
-		IsDeleted:    index.IsDeleted,
+		TenantID:        tenantID,
+		CollectionID:    index.CollectionID,
+		FieldID:         index.FieldID,
+		IndexID:         index.IndexID,
+		IndexName:       index.IndexName,
+		TypeParams:      string(typeParamsBytes),
+		IndexParams:     string(indexParamsBytes),
+		CreateTime:      index.CreateTime,
+		IsDeleted:       index.IsDeleted,
+		IsAutoIndex:     index.IsAutoIndex,
+		UserIndexParams: string(userIndexParamsBytes),
 	}
 
 	err = tc.metaDomain.IndexDb(ctx).Insert([]*dbmodel.Index{idx})
