+++ tests/python_client/testcases/test_connection.py
@@ -542,7 +542,7 @@ class TestConnectionOperation(TestcaseBase):
                                                  check_items={ct.dict_content: {"address": f"{host}:{port}",
                                                                                 "user": ""}})
 
-    @pytest.mark.tags(ct.CaseLabel.L1)
+    @pytest.mark.tags(ct.CaseLabel.L2)
     @pytest.mark.parametrize("connect_name", [DefaultConfig.DEFAULT_USING, "test_alias_nme"])
     def test_connection_connect_wrong_params(self, host, port, connect_name):
         """
